feat: move functions to utils
